Interim, restructuring to install as lib
