+++ .github/workflows/build.yml
@@ -3,6 +3,7 @@ name: Build and Test
 on:
   push:
     branches: [ main, develop ]
+    tags: [ 'v*' ]  # Add this to trigger on tags
   pull_request:
     branches: [ main ]
 
@@ -15,18 +16,18 @@ jobs:
 
     steps:
     - uses: actions/checkout@v4
-    
+
     - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v4
       with:
         python-version: ${{ matrix.python-version }}
-    
+
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
         pip install -e .
         pip install pytest pytest-asyncio httpx
-    
+
     - name: Run tests
       run: pytest tests/ -v
 
@@ -35,44 +36,42 @@ jobs:
     needs: test
     steps:
     - uses: actions/checkout@v4
-    
+
     - name: Set up Python
       uses: actions/setup-python@v4
       with:
         python-version: "3.11"
-    
+
     - name: Install build dependencies
       run: |
         python -m pip install --upgrade pip
         pip install build twine
-    
+
     - name: Build package
       run: python -m build
-    
+
     - name: Check package
       run: twine check dist/*
-    
+
     - name: Upload build artifacts
       uses: actions/upload-artifact@v4
       with:
-        name: dist-${{ matrix.runs-on }}
-        path: dist-${{ matrix.runs-on }}/
+        name: dist-files
+        path: dist/
 
   release:
     runs-on: ubuntu-latest
     needs: build
     if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags/v')
     steps:
-    - uses: actions/checkout@v4
-    
     - name: Download artifacts
       uses: actions/download-artifact@v4
       with:
-        pattern: dist-*
-        merge-multiple: true
+        name: dist-files
         path: dist/
-    
-    - name: Publish to PyPI
+
+    - name: Publish to TestPyPI
       uses: pypa/gh-action-pypi-publish@release/v1
       with:
-        password: ${{ secrets.PYPI_API_TOKEN }}
+        password: ${{ secrets.TEST_PYPI_API_TOKEN }}
+        repository-url: https://test.pypi.org/legacy/
