+++ src/sauce_api_mcp/__main__.py
@@ -1,0 +1,6 @@
+"""Entry point for running sauce-api-mcp as a module."""
+
+from .main import main
+
+if __name__ == "__main__":
+    main()
